requirements.txt
langchain_core
streamlit
langchain
langchain_openai
python-dotenv